+++ topothehourbot/utilities/channels/__init__.py
@@ -1,3 +1,3 @@
-from .builtins import *
+from .core import *
 from .protocols import *
 from .series import *
